+++ src/Database/Models/Group.php
@@ -3,24 +3,21 @@
 /*
  * UF Custom User Profile Field Sprinkle
  *
- * @link https://github.com/lcharette/UF_UserProfile
- * @copyright Copyright (c) 2017 Louis Charette
- * @license https://github.com/lcharette/UF_UserProfile/blob/master/LICENSE (MIT License)
+ * @link      https://github.com/lcharette/UF_UserProfile
+ * @copyright Copyright (c) 2020 Louis Charette
+ * @license   https://github.com/lcharette/UF_UserProfile/blob/master/LICENSE (MIT License)
  */
 
 namespace UserFrosting\Sprinkle\UserProfile\Database\Models;
 
 use UserFrosting\Sprinkle\Account\Database\Models\Group as CoreGroup;
+use UserFrosting\Sprinkle\Core\Database\Relations\MorphManySyncable;
 use UserFrosting\Sprinkle\UserProfile\Database\Models\Traits\ProfileFieldsHelpers;
 
 /**
  * Group Class.
  *
  * Extend the core Group Model to add the custom group profile fields
- *
- * @extend CoreGroup
- *
- * @author Louis Charette
  */
 class Group extends CoreGroup
 {
@@ -28,6 +25,8 @@ class Group extends CoreGroup
 
     /**
      * Eloquent relation to the profile table.
+     *
+     * @return MorphManySyncable
      */
     public function profileFields()
     {
@@ -36,8 +35,7 @@ class Group extends CoreGroup
 
     /**
      * Delete the profileFields values when deleting the main model.
-     *
-     * @return void
+     * {@inheritdoc}
      */
     public function delete()
     {