+++ src/ServicesProvider/ServicesProvider.php
@@ -3,33 +3,37 @@
 /*
  * UF Custom User Profile Field Sprinkle
  *
- * @link https://github.com/lcharette/UF_UserProfile
- * @copyright Copyright (c) 2017 Louis Charette
- * @license https://github.com/lcharette/UF_UserProfile/blob/master/LICENSE (MIT License)
+ * @link      https://github.com/lcharette/UF_UserProfile
+ * @copyright Copyright (c) 2020 Louis Charette
+ * @license   https://github.com/lcharette/UF_UserProfile/blob/master/LICENSE (MIT License)
  */
 
 namespace UserFrosting\Sprinkle\UserProfile\ServicesProvider;
 
+use Psr\Container\ContainerInterface;
+use UserFrosting\Sprinkle\UserProfile\Database\Models\Group;
+use UserFrosting\Sprinkle\UserProfile\Database\Models\User;
+
 /**
- * @author Alex Weissman (https://alexanderweissman.com)
+ * Registers services for the UserProfile sprinkle.
  */
 class ServicesProvider
 {
     /**
-     * Register extended user fields services.
+     * Register UserProfile services.
      *
-     * @param Container $container A DI container implementing ArrayAccess and container-interop.
+     * @param ContainerInterface $container
      */
     public function register($container)
     {
         /*
          * Extend the 'classMapper' service to register model classes.
          *
-         * Mappings added: OwlerUser
+         * Mappings added: user & group
          */
         $container->extend('classMapper', function ($classMapper, $c) {
-            $classMapper->setClassMapping('user', 'UserFrosting\Sprinkle\UserProfile\Database\Models\User');
-            $classMapper->setClassMapping('group', 'UserFrosting\Sprinkle\UserProfile\Database\Models\Group');
+            $classMapper->setClassMapping('user', User::class);
+            $classMapper->setClassMapping('group', Group::class);
 
             return $classMapper;
         });