+++ src/UserProfile.php
@@ -3,9 +3,9 @@
 /*
  * UF Custom User Profile Field Sprinkle
  *
- * @link https://github.com/lcharette/UF_UserProfile
- * @copyright Copyright (c) 2017 Louis Charette
- * @license https://github.com/lcharette/UF_UserProfile/blob/master/LICENSE (MIT License)
+ * @link      https://github.com/lcharette/UF_UserProfile
+ * @copyright Copyright (c) 2020 Louis Charette
+ * @license   https://github.com/lcharette/UF_UserProfile/blob/master/LICENSE (MIT License)
  */
 
 namespace UserFrosting\Sprinkle\UserProfile;
@@ -13,11 +13,7 @@ namespace UserFrosting\Sprinkle\UserProfile;
 use UserFrosting\System\Sprinkle\Sprinkle;
 
 /**
- * UserProfile class.
- *
  * Bootstrapper class for the 'UserProfile' sprinkle.
- *
- * @extends Sprinkle
  */
 class UserProfile extends Sprinkle
 {